Managing state and life cycle
In this tutorial, you'll use reference type models with SwiftUI views.
You'll display meeting time and progress in views that better represent
the final design.

I was supposed to learn how to integrate AVFoundation to play audio
when the scrum's state changes, but when I tried to follow along with
the tutorial I got weird code signing issues and the app could no longer
build so I opted to leave that feature out when following along with the
tutorial.